# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/overview:
    get:
      operationId: getAdminOverview
      tags: [admin]
      summary: Get protected admin overview
      responses:
        "200":
          description: Admin overview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminOverview"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /admin/users:
    get:
      operationId: listAdminUsers
      tags: [admin]
      summary: List users for administration
      responses:
        "200":
          description: Users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminUser"
        "403":
          $ref: "#/components/responses/Forbidden"
  /admin/users/{userId}/role:
    patch:
      operationId: updateUserRole
      tags: [admin]
      summary: Grant or revoke admin permission
      parameters:
        - $ref: "#/components/parameters/UserId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoleUpdateInput"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
        "403":
          $ref: "#/components/responses/Forbidden"
  /admin/withdrawals:
    get:
      operationId: listAdminWithdrawals
      tags: [admin]
      summary: List withdrawals awaiting review
      responses:
        "200":
          description: Withdrawal queue
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Withdrawal"
        "403":
          $ref: "#/components/responses/Forbidden"
  /admin/deposits:
    get:
      operationId: listAdminDeposits
      tags: [admin]
      summary: List deposits awaiting verification
      responses:
        "200":
          description: Deposit queue
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminDeposit"
        "403":
          $ref: "#/components/responses/Forbidden"
  /admin/deposits/{depositId}/review:
    post:
      operationId: reviewDeposit
      tags: [admin]
      summary: Approve or reject a submitted deposit
      parameters:
        - $ref: "#/components/parameters/DepositId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepositReviewInput"
      responses:
        "200":
          description: Reviewed deposit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminDeposit"
        "403":
          $ref: "#/components/responses/Forbidden"
  /admin/withdrawals/{withdrawalId}/review:
    post:
      operationId: reviewWithdrawal
      tags: [admin]
      summary: Approve or reject a withdrawal
      parameters:
        - $ref: "#/components/parameters/WithdrawalId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WithdrawalReviewInput"
      responses:
        "200":
          description: Reviewed withdrawal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Withdrawal"
        "403":
          $ref: "#/components/responses/Forbidden"
components:
  parameters:
...
    WithdrawalId:
      name: withdrawalId
      in: path
      required: true
      schema:
        type: string
    DepositId:
      name: depositId
      in: path
      required: true
      schema:
        type: string
    UserId:
      name: userId
      in: path
      required: true
      schema:
        type: string
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Insufficient permission
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]
    UserProfile:
      type: object
      required: [id, name, email, role, balance, referralCode, depositReference, createdAt]
      properties:
        id: { type: string }
        name: { type: string }
        email: { type: string, format: email }
        avatarUrl: { type: ["string", "null"], format: uri }
        role:
          type: string
          enum: [worker, advertiser, admin, super_admin]
        balance: { type: number }
        referralCode: { type: string }
        depositReference: { type: string }
        referredBy: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
...
    Withdrawal:
      type: object
      required: [id, amount, method, status, requestedAt]
      properties:
        id: { type: string }
        amount: { type: number }
        method: { type: string, enum: [bank_transfer, mobile_money, crypto] }
        accountLabel: { type: string }
        status: { type: string, enum: [pending, approved, rejected, paid] }
        note: { type: ["string", "null"] }
        requestedAt: { type: string, format: date-time }
        processedAt: { type: ["string", "null"], format: date-time }
...
    WithdrawalReviewInput:
      type: object
      required: [decision]
      properties:
        decision: { type: string, enum: [approved, rejected, paid] }
        note: { type: string, maxLength: 500 }
    Deposit:
      type: object
      required: [id, amount, bankName, transferReference, depositReference, transferredAt, status, createdAt]
      properties:
        id: { type: string }
        amount: { type: number }
        bankName: { type: string }
        transferReference: { type: string }
        depositReference: { type: string }
        transferredAt: { type: string, format: date-time }
        proofUrl: { type: ["string", "null"], format: uri }
        status: { type: string, enum: [pending, approved, rejected] }
        note: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
        reviewedAt: { type: ["string", "null"], format: date-time }
...
    DepositReviewInput:
      type: object
      required: [decision]
      properties:
        decision: { type: string, enum: [approved, rejected] }
        note: { type: string, maxLength: 500 }
    AdminDeposit:
      allOf:
        - $ref: "#/components/schemas/Deposit"
        - type: object
          required: [userId, userName, userEmail]
          properties:
            userId: { type: string }
            userName: { type: string }
            userEmail: { type: string, format: email }
    AdminOverview:
      type: object
      required: [totalUsers, activeTasks, pendingSubmissions, pendingWithdrawals, totalVolume, platformRevenue, payoutSplit]
      properties:
        totalUsers: { type: integer }
        activeTasks: { type: integer }
        pendingSubmissions: { type: integer }
        pendingWithdrawals: { type: integer }
        totalVolume: { type: number }
        platformRevenue: { type: number }
        payoutSplit:
          type: object
          required: [owner, worker, referrer]
          properties:
            owner: { type: number }
            worker: { type: number }
            referrer: { type: number }
    AdminUser:
      allOf:
        - $ref: "#/components/schemas/UserProfile"
        - type: object
          required: [lastActiveAt]
          properties:
            lastActiveAt: { type: string, format: date-time }
    RoleUpdateInput:
      type: object
      required: [role]
      properties:
        role: { type: string, enum: [worker, advertiser, admin] }