# lib/api-zod/src/generated/types/deposit.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DepositStatus } from './depositStatus';

export interface Deposit {
  id: string;
  amount: number;
  bankName: string;
  transferReference: string;
  depositReference: string;
  transferredAt: Date;
  /** @nullable */
  proofUrl?: string | null;
  status: DepositStatus;
  /** @nullable */
  note?: string | null;
  createdAt: Date;
  /** @nullable */
  reviewedAt?: Date | null;
}