# lib/api-zod/src/generated/types/depositReviewInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DepositReviewInputDecision } from './depositReviewInputDecision';

export interface DepositReviewInput {
  decision: DepositReviewInputDecision;
  /** @maxLength 500 */
  note?: string;
}