# lib/api-zod/src/generated/types/userProfile.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UserProfileRole } from './userProfileRole';

export interface UserProfile {
  id: string;
  name: string;
  email: string;
  /** @nullable */
  avatarUrl?: string | null;
  role: UserProfileRole;
  balance: number;
  referralCode: string;
  depositReference: string;
  /** @nullable */
  referredBy?: string | null;
  createdAt: Date;
}